requirements.txt
ffmpeg-python
numpy
pyyaml
boto3
botocore